Big epic refactor of everything.
Decided to test my boundaries by rewriting the whole
thing in one-fell-swoop with zero tests or feedback.

Went pretty well. Pressures that led me to this
change were mostly the gigantic function-bag of 33
loosely-related methods that had been previously
driving all the behavior. Because the supporting
code really depended on a huge surface area of
methods to be within reach, all of that had to be
rethought too.

A couple fun things:

* Split the provided Warrior class into a Brain
and Body wrapper. This lets me delegate questions
to the brain and only take actions against the body
* The body wrapper isolates the game's restricition
of "only one action" by turn by effectively ignoring
subsequent actions taken. This basically saved me
an "else" condition by anyone who'd conceivably
want to take an action, as now it's just a no-op.